[DF-14435] Fixed condition value, changed expected value in test case
Former-commit-id: d708162c7e2f9d8e8aec6314d9f8371252707490

+++ Sources/Internal/Render/RHI/Common/PreProcessor.cpp
@@ -564,7 +564,7 @@ bool PreProc::ProcessBuffer(char* inputText, LineVector& lines)
                     *s = NewLine; // since it was null'ed in GetExpression
                     ln = s + 1;
 
-                    bool condition = static_cast<int32>(v) != 0;
+                    bool condition = (v != 0.0f);
 
                     condition_t p;
                     p.original_condition = condition;
